+++ src/extension.ts
@@ -8,26 +8,9 @@
 
 import * as vscode from 'vscode';
 import * as engine from './engine';
-import crypto from 'node:crypto';
+import { $$logErrors } from './globals';
 
-/* ******************************************************
- *
- * to modify text in VSCode:
- *
- * https://github.com/Microsoft/vscode-extension-samples/tree/main/document-editing-sample
- *
- * a library that abstracts upon crypto
- *
- * https://www.npmjs.com/package/crypto-shield?activeTab=readme
- *
- */
-
-// ********************
-// CODICE DI PROVA. VA CANCELLATO PRIMA DEL RILASCIO
-//
-const pw = 'sammarino';
-
-const applyToCurrentFile = (operation: Function) => {
+const applyToCurrentFile = (operation: Function, key: string) => {
   const editor = vscode.window.activeTextEditor;
   if (!editor) return;
   const text = editor.document.getText();
@@ -43,27 +26,75 @@ const applyToCurrentFile = (operation: Function) => {
     new vscode.Position(0, 0),
     new vscode.Position(lastLine, lastChar)
   );
-  operation(text, pw).then((encrText: string) => {
+  operation(text, key).then((encrText: string) => {
     editor.edit((editBuilder) => {
       editBuilder.delete(editor.selection);
       editBuilder.insert(new vscode.Position(0, 0), encrText);
     });
   });
 };
-// *********************
+
+const getUserKey = async (context: 'encryption' | 'decryption') => {
+  const firstKey = await vscode.window.showInputBox({
+    password: true,
+    placeHolder: 'Your secret key',
+    prompt: 'Type a key',
+    title: `File ${context}`,
+  });
+
+  if (!firstKey) {
+    if ($$logErrors) {
+      console.log(
+        '§> getUserKey: There is no key or the key is invalid. Procedure aborted.'
+      );
+      return;
+    }
+  }
+
+  const secondKey = await vscode.window.showInputBox({
+    password: true,
+    placeHolder: 'Your secret key',
+    prompt: 'Confirm your key',
+    title: `File ${context}`,
+  });
+
+  if (!secondKey) {
+    if ($$logErrors) {
+      console.log(
+        '§> getUserKey: There is no key or the key is invalid. Procedure aborted.'
+      );
+      return;
+    }
+  }
+
+  if (firstKey === secondKey) {
+    return firstKey;
+  } else {
+    console.log('§>', 'Key mismatch. Procedure aborted');
+    return;
+  }
+};
 
 export function activate(context: vscode.ExtensionContext) {
   let disposableEncrypt = vscode.commands.registerCommand(
     'sams-file-encryptor.encrypt',
     () => {
-      applyToCurrentFile(engine.encrypt);
+      getUserKey('encryption').then((key) => {
+        if (key) {
+          applyToCurrentFile(engine.encrypt, key);
+        }
+      });
     }
   );
 
   let disposableDecrypt = vscode.commands.registerCommand(
     'sams-file-encryptor.decrypt',
     () => {
-      applyToCurrentFile(engine.decrypt);
+      getUserKey('decryption').then((key) => {
+        if (key) {
+          applyToCurrentFile(engine.decrypt, key);
+        }
+      });
     }
   );