+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Change Log
 
 <!--
-## [major.minor.patch] - yyyy-mm-dd
+## [Unreleased] | [major.minor.patch] - yyyy-mm-dd
 ### Added | Fixed | Changed | Removed | Deprecated | Security
 - filename {section}: description
 -->